feat: exchange module package
