+++ libs/exchange-module/.eslintignore
@@ -1,0 +1,2 @@
+lib/
+lib-es/
