+++ libs/exchange-module/.unimportedrc.json
@@ -1,0 +1,5 @@
+{
+  "entry": ["src/index.ts"],
+  "ignoreUnimported": [],
+  "ignoreUnresolved": []
+}
