+++ libs/exchange-module/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "jsdom",
+  testPathIgnorePatterns: ["lib/", "lib-es/"],
+  globals: {
+    "ts-jest": {
+      isolatedModules: true,
+    },
+  },
+};
