+++ libs/exchange-module/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../../tsconfig.base",
+  "compilerOptions": {
+    "declaration": true,
+    "declarationMap": true,
+    "noImplicitAny": false,
+    "noImplicitThis": false,
+    "downlevelIteration": true,
+    "module": "commonjs",
+    "lib": ["es2020", "dom"],
+    "jsx": "react",
+    "outDir": "lib"
+  },
+  "include": ["src/**/*"]
+}
